Added Main Menu, cleaned up panels
Now the panels make more sense, and the main menu works

+++ Tower Defense/GameProject/src/entities/Critter_Arrow.java
@@ -17,7 +17,7 @@ public class Critter_Arrow extends Critter{
 		//does not regenerate health
 		regen = 0;
 		//fast
-		speed = (int) (5*levelMultiplier);
+		speed = (5*levelMultiplier);
 		//does not resist effects
 		resistance = 0;
 		//set name
